Top framse now in stack
Instead of keeping single frame with vname -> id mappings
we use entire list to allow nested scopes.

Even if we do not allow to shadow variables names,
it is required to allow loops with declarations.

If we allow on shadowing, then we need to adjust
`Intermediate.instantiateFunction` to take values from right frame.